refactor: move PostgreSQL-specific code to sqldb-postgresql module
Reorganize source code to separate database-agnostic and database-specific
implementations following the module structure created in Phase 1.

Changes:
- Move MessageCatalogue entity to PostgreSQL module (uses @Table(PostgresDbType))
- Move MessageCatalogueRepository implementation to PostgreSQL module
- Move all PostgreSQL infrastructure (Transactor, DataSource, Config, DatabaseSupport)
- Move migration tools to sqldb-postgresql/migration package
- Move SQL migration files to database-specific directories:
  * V1__create_message_catalogue.postgresql.sql → sqldb-postgresql/src/main/resources/db/migration/postgresql/V1__create_message_catalogue.sql
  * V1__create_message_catalogue.mysql.sql → sqldb-mysql/src/main/resources/db/migration/mysql/V1__create_message_catalogue.sql
- Move all tests and test resources to PostgreSQL module
- Split FlywayMigrationService: keep trait/config in core, move PostgreSQL impl to sqldb-postgresql
- Update package declarations to works.iterative.sqldb.postgresql

The core sqldb module now only contains database-agnostic code:
- FlywayMigrationService trait
- FlywayConfig

IWSD-76
