+++ sqldb-mysql/src/main/resources/db/migration/mysql/V1__create_message_catalogue.sql
