+++ sqldb-postgresql/src/main/resources/db/migration/postgresql/V1__create_message_catalogue.sql
