+++ sqldb-postgresql/src/main/scala/works/iterative/sqldb/postgresql/PostgreSQLConfig.scala
@@ -1,4 +1,4 @@
-package works.iterative.sqldb
+package works.iterative.sqldb.postgresql
 import zio.*
 
 /** Configuration for PostgreSQL database connection
