+++ sqldb-postgresql/src/main/scala/works/iterative/sqldb/postgresql/PostgreSQLDataSource.scala
@@ -1,4 +1,4 @@
-package works.iterative.sqldb
+package works.iterative.sqldb.postgresql
 
 import zio.*
 import javax.sql.DataSource
